only admin action protected

+++ app/Http/Controllers/ActivityController.php
@@ -7,14 +7,22 @@ use App\Models\Activity;
 
 class ActivityController extends Controller
 {
+    /**
+     * Display a listing of activities.
+     *
+     * @return \Illuminate\Http\Response
+     */
     public function show($name)
     {
         return Activity::where('name',$name)->first();
-        //return $user;
     }
 
     public function store(Request $request){
-        //return response()->json("Activity created");
+
+        if($request->user()->role != 'admin'){
+            return response()->json(['message' => 'Unauthorized'], 401);
+        }
+
         Activity::create($request->validate([ 
             'name' => ['required', 'unique:activities'],
             'weight' => ['required'],
@@ -25,6 +33,9 @@ class ActivityController extends Controller
 
     public function update(Request $request, $activityName)
     {
+        if($request->user()->role != 'admin'){
+            return response()->json(['message' => 'Unauthorized'], 401);
+        }
         $activity = Activity::where('name',$activityName)->first();
         $activity->update($request->validate([ 
             'weight' => ['required']
@@ -33,7 +44,10 @@ class ActivityController extends Controller
         return response()->json("Activity updated");
     }
 
-    public function destroy($activityName){
+    public function destroy(Request $request, $activityName){
+        if($request->user()->role != 'admin'){
+            return response()->json(['message' => 'Unauthorized'], 401);
+        }
 
         Activity::where('name',$activityName)->first()->delete();
         return response()->json("Activity " . $activityName . " deleted");
