+++ app/Http/Controllers/AttendanceDaysUsersController.php
@@ -9,6 +9,11 @@ use Illuminate\Support\Facades\DB;
 
 class AttendanceDaysUsersController extends Controller
 {
+    /**
+     * Display whole attendance.
+     *
+     * @return \Illuminate\Http\Response
+     */
     public function getAttendanceUsersTable(){
         $users =  User::where('username', '!=', 'admin')->orderBy('username')->get();
         $attendanceDays = AttendanceDay::orderBy('date')->get();
@@ -21,6 +26,10 @@ class AttendanceDaysUsersController extends Controller
     }
 
     public function add(Request $request){
+      if($request->user()->role != 'admin'){
+        return response()->json(['message' => 'Unauthorized'], 401);
+      }
+
         $username = $request->input('username');
         $attendance_day_id = $request->input('attendance_day_id');
   
@@ -29,18 +38,19 @@ class AttendanceDaysUsersController extends Controller
   
         $user->attendance_days()->attach($attendanceDay);
         return response()->json("Attendance created");
+    }
+  
+    public function delete(Request $request){
+      if($request->user()->role != 'admin'){
+        return response()->json(['message' => 'Unauthorized'], 401);
       }
+      $username = $request->input('username');
+      $attendance_day_id = $request->input('attendance_day_id');
   
-      public function delete(Request $request){
-        $username = $request->input('username');
-        $attendance_day_id = $request->input('attendance_day_id');
+      $user =  User::where('username',$username)->first();
+      $attendanceDay = AttendanceDay::find($attendance_day_id);
   
-        $user =  User::where('username',$username)->first();
-        $attendanceDay = AttendanceDay::find($attendance_day_id);
-  
-        $user->attendance_days()->detach($attendanceDay);
-        return response()->json("Activity marked as not done");
-      }
-
-
+      $user->attendance_days()->detach($attendanceDay);
+      return response()->json("Activity marked as not done");
+    }
 }
